# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: UniTeam API
  description: Gestión colaborativa de tareas para equipos universitarios.
  version: "0.2.0"
servers:
  - url: https://uniteam-api.onrender.com
    description: Producción (Render). URL definitiva en docs/despliegue/guia.md
  - url: http://localhost:8000
    description: Servidor local de desarrollo
paths:
  /:
    get:
      summary: Estado del servicio
      operationId: raiz
      tags:
        - sistema
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /activo:
    get:
      summary: Estado del servicio
      operationId: activo
      tags:
        - sistema
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
  /health:
    get:
      summary: Health check con comprobación de la base de datos
      description: >
        200 si el proceso responde y alcanza la base de datos; 503 si no. Lo
        consulta Render para decidir si una instancia recibe tráfico. No exige
        credencial.
      operationId: salud
      tags:
        - sistema
      responses:
        "200":
          description: Disponible
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Salud"
        "503":
          description: La base de datos no responde
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Salud"
  /metricas:
    get:
      summary: Métricas en formato de exposición de Prometheus
      description: >
        Incluye uniteam_tablero_latencia_segundos (ESC-01) y
        uniteam_accesos_denegados_total (ESC-03). Las rutas aparecen como
        plantilla, sin identificadores ni datos personales.
      operationId: metricas
      tags:
        - sistema
      responses:
        "200":
          description: Texto en formato Prometheus 0.0.4
          content:
            text/plain:
              schema:
                type: string
  /metricas/esc-01:
    get:
      summary: Latencia del tablero frente al umbral de ESC-01
      description: >
        Percentiles de las últimas 1000 consultas del tablero atendidas con
        200, medidos dentro de la API, y si cumplen p95 ≤ 2 s y p99 ≤ 4 s.
      operationId: metrica_esc01
      tags:
        - sistema
      responses:
        "200":
          description: Resumen del escenario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumenESC01"
  /mis-tareas:
    get:
      summary: Tareas asignadas al usuario en todos sus proyectos
      description: >
        Solo proyectos de los que el usuario es miembro. Primero las que tienen
        fecha límite, de la más cercana a la más lejana. Hasta 200. Sin
        `incluir_terminadas`, omite las que están en el estado final.
      operationId: mis_tareas
      tags:
        - tareas
      security:
        - BearerAuth: []
      parameters:
        - name: incluir_terminadas
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Tareas asignadas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MiTareaSalida"
        "401":
          $ref: "#/components/responses/NoAutenticado"
  /flujo-estados:
    get:
      summary: Flujo de estados de las tareas
      description: >
        Estados en el orden de las columnas del tablero, con su etiqueta, si son
        el inicial o el final, y a qué estados se puede pasar desde cada uno. Lo
        define el dominio (ADR 0012); la Aplicación Web no conoce los estados de
        otra forma. No exige credencial.
      operationId: flujo_estados
      tags:
        - tareas
      responses:
        "200":
          description: Flujo de estados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlujoEstados"
  /proyectos:
    get:
      summary: Listar proyectos propios
      operationId: listar_mis_proyectos
      tags:
        - proyectos
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Proyectos de los que el usuario es miembro
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProyectoConResumen"
        "401":
          $ref: "#/components/responses/NoAutenticado"
    post:
      summary: Crear proyecto
      operationId: crear_proyecto
      tags:
        - proyectos
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrearProyecto"
      responses:
        "201":
          description: Proyecto creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProyectoSalida"
        "400":
          $ref: "#/components/responses/SolicitudIncorrecta"
        "401":
          $ref: "#/components/responses/NoAutenticado"
  /proyectos/{proyecto_id}:
    get:
      summary: Obtener detalle de proyecto
      operationId: obtener_proyecto
      tags:
        - proyectos
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/proyecto_id"
      responses:
        "200":
          description: Detalle del proyecto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProyectoDetalle"
        "401":
          $ref: "#/components/responses/NoAutenticado"
        "403":
          $ref: "#/components/responses/Prohibido"
  /proyectos/{proyecto_id}/miembros:
    post:
      summary: Agregar miembro al proyecto
      operationId: agregar_miembro
      tags:
        - proyectos
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/proyecto_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgregarMiembro"
      responses:
        "201":
          description: Miembro agregado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProyectoDetalle"
        "400":
          $ref: "#/components/responses/SolicitudIncorrecta"
        "401":
          $ref: "#/components/responses/NoAutenticado"
        "403":
          $ref: "#/components/responses/Prohibido"
        "409":
          $ref: "#/components/responses/Conflicto"
  /proyectos/{proyecto_id}/tareas:
    get:
      summary: Listar tareas (tablero)
      operationId: consultar_tablero
      tags:
        - tareas
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/proyecto_id"
        - name: estado
          in: query
          schema:
            type: string
            enum: [pendiente, en_progreso, en_revision, completada]
        - name: responsable
          in: query
          schema:
            type: string
        - name: limite
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
          default: 50
        - name: desplazamiento
          in: query
          schema:
            type: integer
            minimum: 0
          default: 0
      responses:
        "200":
          description: Lista de tareas paginada
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TareaSalida"
        "401":
          $ref: "#/components/responses/NoAutenticado"
        "403":
          $ref: "#/components/responses/Prohibido"
        "422":
          $ref: "#/components/responses/NoProcesable"
    post:
      summary: Crear tarea
      operationId: crear_tarea
      tags:
        - tareas
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/proyecto_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrearTarea"
      responses:
        "201":
          description: Tarea creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TareaSalida"
        "400":
          $ref: "#/components/responses/SolicitudIncorrecta"
        "401":
          $ref: "#/components/responses/NoAutenticado"
        "403":
          $ref: "#/components/responses/Prohibido"
  /proyectos/{proyecto_id}/tareas/{tarea_id}:
    get:
      summary: Obtener tarea
      operationId: obtener_tarea
      tags:
        - tareas
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/proyecto_id"
        - $ref: "#/components/parameters/tarea_id"
      responses:
        "200":
          description: Detalle de la tarea
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TareaSalida"
        "401":
          $ref: "#/components/responses/NoAutenticado"
        "403":
          $ref: "#/components/responses/Prohibido"
        "404":
          $ref: "#/components/responses/NoEncontrado"
    patch:
      summary: Editar tarea
      description: >
        Edición parcial de título, prioridad y fecha límite. Un campo ausente
        no cambia; `fecha_limite: null` la quita. Estado y responsable tienen
        sus propias rutas. Cualquier miembro del proyecto puede editar.
      operationId: editar_tarea
      tags:
        - tareas
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/proyecto_id"
        - $ref: "#/components/parameters/tarea_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditarTarea"
      responses:
        "200":
          description: Tarea editada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TareaSalida"
        "401":
          $ref: "#/components/responses/NoAutenticado"
        "403":
          $ref: "#/components/responses/Prohibido"
        "404":
          $ref: "#/components/responses/NoEncontrado"
        "422":
          $ref: "#/components/responses/NoProcesable"
    delete:
      summary: Eliminar tarea
      description: >
        Solo quien creó la tarea o el líder del proyecto. El intento de otro
        miembro responde 403 y queda auditado.
      operationId: eliminar_tarea
      tags:
        - tareas
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/proyecto_id"
        - $ref: "#/components/parameters/tarea_id"
      responses:
        "204":
          description: Tarea eliminada
        "401":
          $ref: "#/components/responses/NoAutenticado"
        "403":
          $ref: "#/components/responses/Prohibido"
        "404":
          $ref: "#/components/responses/NoEncontrado"
  /proyectos/{proyecto_id}/tareas/{tarea_id}/responsable:
    put:
      summary: Asignar responsable
      operationId: asignar_tarea
      tags:
        - tareas
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/proyecto_id"
        - $ref: "#/components/parameters/tarea_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AsignarTarea"
      responses:
        "200":
          description: Tarea actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TareaSalida"
        "401":
          $ref: "#/components/responses/NoAutenticado"
        "403":
          $ref: "#/components/responses/Prohibido"
  /proyectos/{proyecto_id}/tareas/{tarea_id}/estado:
    put:
      summary: Cambiar estado
      operationId: cambiar_estado
      tags:
        - tareas
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/proyecto_id"
        - $ref: "#/components/parameters/tarea_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CambiarEstado"
      responses:
        "200":
          description: Tarea actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TareaSalida"
        "401":
          $ref: "#/components/responses/NoAutenticado"
        "403":
          $ref: "#/components/responses/Prohibido"
        "409":
          $ref: "#/components/responses/Conflicto"
  /proyectos/{proyecto_id}/progreso:
    get:
      summary: Consultar progreso
      operationId: consultar_progreso
      tags:
        - proyectos
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/proyecto_id"
      responses:
        "200":
          description: Resumen del avance del proyecto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgresoSalida"
        "401":
          $ref: "#/components/responses/NoAutenticado"
        "403":
          $ref: "#/components/responses/Prohibido"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    proyecto_id:
      name: proyecto_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    tarea_id:
      name: tarea_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    MiTareaSalida:
      allOf:
        - $ref: "#/components/schemas/TareaSalida"
        - type: object
          required: [proyecto_nombre]
          properties:
            proyecto_nombre:
              type: string
    ProyectoConResumen:
      description: >
        Proyecto de la lista con su avance. `resumen` se añadió el 2026-09-28;
        los campos anteriores no cambian.
      allOf:
        - $ref: "#/components/schemas/ProyectoDetalle"
        - type: object
          required: [resumen]
          properties:
            resumen:
              $ref: "#/components/schemas/ResumenProyecto"
    ResumenProyecto:
      type: object
      required: [total, terminadas, vencidas]
      properties:
        total:
          type: integer
        terminadas:
          type: integer
        vencidas:
          type: integer
    FlujoEstados:
      type: object
      required: [estados]
      properties:
        estados:
          type: array
          items:
            $ref: "#/components/schemas/EstadoFlujo"
    EstadoFlujo:
      type: object
      required: [id, etiqueta, inicial, final, siguientes]
      properties:
        id:
          type: string
          example: pendiente
        etiqueta:
          type: string
          example: Pendiente
        inicial:
          type: boolean
        final:
          type: boolean
        siguientes:
          type: array
          items:
            type: string
    Salud:
      type: object
      required: [estado, base_de_datos, version, revision, entorno]
      properties:
        estado:
          type: string
          enum: [ok, degradado]
        base_de_datos:
          type: string
          enum: [ok, caida]
        latencia_bd_ms:
          type: number
        version:
          type: string
        revision:
          type: string
          description: Commit desplegado (12 caracteres)
        entorno:
          type: string
    ResumenESC01:
      type: object
      required: [escenario, umbral_p95_s, umbral_p99_s, muestras, cumple]
      properties:
        escenario:
          type: string
          example: ESC-01
        descripcion:
          type: string
        umbral_p95_s:
          type: number
          example: 2.0
        umbral_p99_s:
          type: number
          example: 4.0
        ventana:
          type: integer
        muestras:
          type: integer
        p50_s:
          type: number
          nullable: true
        p95_s:
          type: number
          nullable: true
        p99_s:
          type: number
          nullable: true
        cumple:
          type: boolean
          nullable: true
          description: null mientras no haya muestras
        alcance:
          type: string
    CrearProyecto:
      type: object
      required: [nombre]
      properties:
        nombre:
          type: string
          minLength: 1
          maxLength: 200
        miembros:
          type: array
          items:
            type: string
          default: []
    AgregarMiembro:
      type: object
      required: [usuario]
      properties:
        usuario:
          type: string
          minLength: 1
          maxLength: 120
        rol:
          type: string
          enum: [integrante, lider]
          default: integrante
    CrearTarea:
      type: object
      required: [titulo]
      properties:
        titulo:
          type: string
          minLength: 1
          maxLength: 300
        prioridad:
          type: string
          enum: [baja, media, alta]
          default: media
        responsable:
          type: string
          nullable: true
        fecha_limite:
          type: string
          format: date
          nullable: true
    AsignarTarea:
      type: object
      required: [responsable]
      properties:
        responsable:
          type: string
    EditarTarea:
      type: object
      minProperties: 1
      description: Al menos un campo. Los identificadores de usuario se normalizan a minúsculas.
      properties:
        titulo:
          type: string
          minLength: 1
          maxLength: 300
        prioridad:
          type: string
          enum: [baja, media, alta]
        fecha_limite:
          type: string
          format: date
          nullable: true
          description: null quita la fecha límite
    CambiarEstado:
      type: object
      required: [estado]
      properties:
        estado:
          type: string
          enum: [pendiente, en_progreso, en_revision, completada]
    ProyectoSalida:
      type: object
      required: [id, nombre, miembros]
      properties:
        id:
          type: string
        nombre:
          type: string
        miembros:
          type: array
          items:
            type: string
    ProyectoDetalle:
      type: object
      required: [id, nombre, miembros]
      properties:
        id:
          type: string
        nombre:
          type: string
        miembros:
          type: array
          items:
            $ref: "#/components/schemas/MiembroSalida"
    MiembroSalida:
      type: object
      required: [usuario, rol]
      properties:
        usuario:
          type: string
        rol:
          type: string
          enum: [integrante, lider]
    TareaSalida:
      type: object
      required: [id, proyecto_id, titulo, prioridad, estado, creada_por, creada_en]
      properties:
        id:
          type: string
        proyecto_id:
          type: string
        titulo:
          type: string
        prioridad:
          type: string
          enum: [baja, media, alta]
        estado:
          type: string
          enum: [pendiente, en_progreso, en_revision, completada]
        responsable:
          type: string
          nullable: true
        fecha_limite:
          type: string
          format: date
          nullable: true
        creada_por:
          type: string
        creada_en:
          type: string
          format: date-time
    ProgresoSalida:
      type: object
      required: [total, por_estado, sin_responsable, vencidas, porcentaje_completado]
      properties:
        total:
          type: integer
        por_estado:
          type: object
          additionalProperties:
            type: integer
        sin_responsable:
          type: integer
        vencidas:
          type: integer
        porcentaje_completado:
          type: number
  responses:
    NoAutenticado:
      description: No autenticado
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
          example:
            detail: "Falta la credencial: se espera 'Authorization: Bearer <token>'."
      headers:
        WWW-Authenticate:
          description: Bearer
    Prohibido:
      description: Acceso denegado
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
          example:
            detail: "No tiene acceso a este proyecto o el proyecto no existe."
    NoEncontrado:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
    Conflicto:
      description: Conflicto
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
    SolicitudIncorrecta:
      description: Solicitud incorrecta
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
    NoProcesable:
      description: Validación fallida
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string